Currently Trying to debug Scaling issue with pressure and temperature output

+++ include/DPS.h
@@ -16,10 +16,10 @@ esp_err_t spi_bus_init(spi_device_handle_t *handle_out);
 esp_err_t dps368_read(spi_device_handle_t dev, uint8_t reg, uint8_t *out, size_t len);
 
 //This gets the coefficients required for pressure calc. stores them in a provided array
-void dps368_get_coeff(spi_device_handle_t dev, uint32_t *coeffs);
+void dps368_get_coeff(spi_device_handle_t dev, int32_t *coeffs);
 
-int corrected_pressure(float raw_pressure, float raw_temp, uint32_t *coeffs);
-int corrected_temperature(float raw_temp, uint32_t *coeffs);
+float corrected_pressure(float raw_pressure, float raw_temp, int32_t *coeffs);
+float corrected_temperature(float raw_temp, int32_t *coeffs);
 
 float dps368_get_raw_temp(spi_device_handle_t dev);
